Move msg iter to separate file
