+++ .gitignore
@@ -1,1 +1,3 @@
 target/
+
+notes.md
